Make `@astrojs/markdown-remark` browser-safe
