+++ .changeset/thick-carrots-run.md
@@ -1,0 +1,5 @@
+---
+"@astrojs/markdown-remark": minor
+---
+
+Uses subpath imports to allow the package to be used in browser environments
