+++ packages/markdown/remark/package.json
@@ -16,6 +16,12 @@
     ".": "./dist/index.js",
     "./dist/internal.js": "./dist/internal.js"
   },
+  "imports": {
+    "#import-plugin": {
+      "browser": "./dist/import-plugin-browser.js",
+      "default": "./dist/import-plugin-default.js"
+    }
+  },
   "files": [
     "dist"
   ],
@@ -52,6 +58,7 @@
     "@types/unist": "^3.0.2",
     "astro-scripts": "workspace:*",
     "chai": "^4.3.7",
+    "esbuild": "^0.19.6",
     "mdast-util-mdx-expression": "^2.0.0",
     "mocha": "^10.2.0"
   },
