+++ packages/markdown/remark/src/load-plugins.ts
@@ -1,26 +1,13 @@
-import { resolve as importMetaResolve } from 'import-meta-resolve';
-import path from 'node:path';
-import { pathToFileURL } from 'node:url';
 import type * as unified from 'unified';
+// @ts-ignore Types may not exist if you have not run `pnpm build` first
+import { importPlugin as _importPlugin } from '#import-plugin';
 
-let cwdUrlStr: string | undefined;
-
-async function importPlugin(p: string | unified.Plugin): Promise<unified.Plugin> {
+async function importPlugin(p: string | unified.Plugin<any[], any>) {
 	if (typeof p === 'string') {
-		// Try import from this package first
-		try {
-			const importResult = await import(p);
-			return importResult.default;
-		} catch {}
-
-		// Try import from user project
-		cwdUrlStr ??= pathToFileURL(path.join(process.cwd(), 'package.json')).toString();
-		const resolved = importMetaResolve(p, cwdUrlStr);
-		const importResult = await import(resolved);
-		return importResult.default;
+		return await _importPlugin(p);
+	} else {
+		return p;
 	}
-
-	return p;
 }
 
 export function loadPlugins(
